refactor: reorganize SQL text editor extensions
- Refactor SQL autocomplete into two extensions (customSql and sqlContextComplete)
- Replace SqlContextAnalyzer with inline functions in sqlContextComplete
- Add support for more SQL keywords in autocomplete trigger
- Improve component type consistency and imports
